+++ lib/elixir/src/elixir_translator.erl
@@ -62,9 +62,6 @@ translate_each({ '__block__', Meta, Args }, S) when is_list(Args) ->
   { TArgs, NS } = translate(Args, S),
   { { block, ?line(Meta), TArgs }, NS };
 
-translate_each({ '__scope__', _Meta, [[{erl,true}],[{do,Expr}]] }, S) ->
-  { Expr, S };
-
 translate_each({ '__scope__', _Meta, [[{file,File}],[{do,Expr}]] }, S) ->
   Old = S#elixir_scope.file,
   { TExpr, TS } = translate_each(Expr, S#elixir_scope{file=File}),
@@ -220,7 +217,7 @@ translate_each({ quote, Meta, [KV, Do] }, S) when is_list(Do) ->
       false -> syntax_error(Meta, S#elixir_scope.file, "invalid args for quote")
     end,
 
-  ValidOpts   = [hygiene, context, var_context, location, line, file, unquote],
+  ValidOpts   = [hygiene, context, var_context, location, line, file, unquote, binding],
   { TKV, ST } = translate_opts(Meta, quote, ValidOpts, KV, S),
 
   Hygiene = case lists:keyfind(hygiene, 1, TKV) of
@@ -261,29 +258,34 @@ translate_each({ quote, Meta, [KV, Do] }, S) when is_list(Do) ->
     false -> DefaultFile
   end,
 
-  TFile = case File of
+  Scope = case File of
     keep -> S#elixir_scope.file;
     _    -> File
   end,
 
-  WExprs = if
-    is_binary(TFile) ->
-      { '__scope__', Meta, [[{file,TFile}],[{do,Exprs}]] };
+  QExprs = if
+    is_binary(Scope) ->
+      { '__scope__', Meta, [[{file,Scope}],[{do,Exprs}]] };
     File == nil ->
       Exprs;
     true ->
       syntax_error(Meta, S#elixir_scope.file, "invalid args for quote, expected :file to be a binary")
   end,
 
+  { Binding, DefaultUnquote } = case lists:keyfind(binding, 1, TKV) of
+    { binding, B } when is_list(B) -> { B, false };
+    false -> { nil, true }
+  end,
+
   Unquote = case lists:keyfind(unquote, 1, TKV) of
-    { unquote, false } -> false;
-    _ -> true
+    { unquote, Bool } when is_boolean(Bool) -> Bool;
+    false -> DefaultUnquote
   end,
 
   Q = #elixir_quote{vars_hygiene=Vars, line=Line, unquote=Unquote,
         aliases_hygiene=Aliases, imports_hygiene=Imports, context=Context},
 
-  { TExprs, _TQ, TS } = elixir_quote:erl_quote(WExprs, Q, ST),
+  { TExprs, _TQ, TS } = elixir_quote:erl_quote(QExprs, Binding, Q, ST),
   { TExprs, TS };
 
 translate_each({ quote, Meta, [_, _] }, S) ->
@@ -473,10 +475,21 @@ translate_each(Literal, S) ->
 %% Opts
 
 translate_opts(Meta, Kind, Allowed, Opts, S) ->
-  { TOpts, TS } = translate_each(Opts, S),
-  FOpts = elixir_tree_helpers:erl_to_elixir(TOpts),
-  validate_opts(Meta, Kind, Allowed, FOpts, TS),
-  { FOpts, TS }.
+  { Expanded, TS } = case literal_opts(Opts) orelse skip_expansion(S#elixir_scope.module) of
+    true  -> { Opts, S };
+    false -> 'Elixir.Macro':expand_all(Opts, elixir_scope:to_ex_env({ ?line(Meta), S }), S)
+  end,
+  validate_opts(Meta, Kind, Allowed, Expanded, TS),
+  { Expanded, TS }.
+
+skip_expansion('Elixir.Kernel') -> true;
+skip_expansion('Elixir.Kernel.Typespec') -> true;
+skip_expansion(_) -> false.
+
+literal_opts({ X, Y }) -> literal_opts(X) andalso literal_opts(Y);
+literal_opts(Opts) when is_list(Opts) -> lists:all(fun literal_opts/1, Opts);
+literal_opts(Opts) when is_atom(Opts); is_number(Opts); is_bitstring(Opts) -> true;
+literal_opts(_) -> false.
 
 validate_opts(Meta, Kind, Allowed, Opts, S) when is_list(Opts) ->
   [begin