+++ lib/elixir/test/elixir/kernel/quote_test.exs
@@ -138,6 +138,13 @@ defmodule Kernel.QuoteTest do
   test :quote_with_dynamic_opts do
     assert quote(dynamic_opts, do: bar(1, 2, 3)) == { :bar, [line: 3], [1, 2, 3] }
   end
+
+  test :binding do
+    assert quote(binding: [foo: 1 + 2], do: foo) == { :__block__, [], [
+      { :=, [], [{ :foo, [], Kernel.QuoteTest }, 3] },
+      { :foo, [], Kernel.QuoteTest }
+    ] }
+  end
 end
 
 defmodule Kernel.QuoteTest.VarHygiene do